third_party/blink/web_tests/virtual/netinfo-constant-type/netinfo/basic-operation-expected.txt

Tests the basic operation of NetInfo.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS navigator.connection is defined.
PASS navigator.connection.type is defined.
PASS navigator.connection.saveData is defined.
PASS navigator.connection.downlinkMax is defined.
PASS navigator.connection.effectiveType is defined.
PASS navigator.connection.rtt is defined.
PASS navigator.connection.downlink is defined.
PASS typeof connection.type is "string"
FAIL connection.type should be ethernet. Was unknown.
PASS connection.saveData is false
FAIL connection.downlinkMax should be 2. Was Infinity.
PASS connection.effectiveType is initialEffectiveType
PASS connection.rtt is initialRtt
PASS connection.downlink is initialDownlink
PASS typeof connection.type is "string"
FAIL connection.type should be bluetooth. Was unknown.
PASS connection.saveData is false
FAIL connection.downlinkMax should be 1. Was Infinity.
PASS connection.effectiveType is initialEffectiveType
PASS connection.rtt is initialRtt
PASS connection.downlink is initialDownlink
PASS successfullyParsed is true

TEST COMPLETE
